refactor: extract MethodBuilder logic out of Method
