+++ src/main/java/com/weirddev/testme/intellij/template/context/Method.java
@@ -1,30 +1,19 @@
 package com.weirddev.testme.intellij.template.context;
 
-import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.util.Pair;
 import com.intellij.psi.*;
-import com.intellij.psi.impl.source.PsiClassReferenceType;
-import com.intellij.psi.search.LocalSearchScope;
-import com.intellij.psi.search.searches.ReferencesSearch;
-import com.intellij.psi.util.MethodSignatureUtil;
-import com.intellij.psi.util.PsiTreeUtil;
-import com.intellij.psi.util.PsiUtil;
-import com.weirddev.testme.intellij.common.utils.LanguageUtils;
+import com.weirddev.testme.intellij.builder.MethodBuilder;
 import com.weirddev.testme.intellij.common.utils.PsiMethodUtils;
-import com.weirddev.testme.intellij.groovy.resolvers.GroovyPsiTreeUtils;
-import com.weirddev.testme.intellij.resolvers.to.MethodCallArg;
-import com.weirddev.testme.intellij.resolvers.to.ResolvedMethodCall;
-import com.weirddev.testme.intellij.resolvers.to.ResolvedReference;
-import com.weirddev.testme.intellij.scala.resolvers.ScalaPsiTreeUtils;
 import com.weirddev.testme.intellij.template.TypeDictionary;
 import com.weirddev.testme.intellij.utils.ClassNameUtils;
-import com.weirddev.testme.intellij.utils.JavaPsiTreeUtils;
 import com.weirddev.testme.intellij.utils.PropertyUtils;
 import lombok.Getter;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.*;
-import java.util.stream.Stream;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Optional;
+import java.util.Set;
 
 /**
  * A class Method.
@@ -32,7 +21,6 @@ import java.util.stream.Stream;
  * @author Yaron Yamin
  */
 public class Method {
-    private static final Logger LOG = Logger.getInstance(Method.class.getName());
     /**
      * Method's return type
      */
@@ -147,6 +135,7 @@ public class Method {
      */
     @Getter private final Set<Field> indirectlyAffectedFields = new HashSet<>();
 
+    //todo need to refactor. use a Factory method to create Method instances
     public Method(PsiMethod psiMethod, PsiClass srcClass, int maxRecursionDepth, TypeDictionary typeDictionary, @Nullable PsiType ownerClassPsiType) {
         isPrivate = psiMethod.hasModifierProperty(PsiModifier.PRIVATE);
         isProtected = psiMethod.hasModifierProperty(PsiModifier.PROTECTED);
@@ -165,44 +154,19 @@ public class Method {
 //        propertyName = psiField == null ? null : psiField.getName();
         propertyName = ClassNameUtils.extractTargetPropertyName(name,isSetter,isGetter);
         if (srcClass != null) {
-            overridden = isOverriddenInChild(psiMethod, srcClass);
-            inherited = isInherited(psiMethod, srcClass);
+            overridden = MethodBuilder.isOverriddenInChild(psiMethod, srcClass);
+            inherited = MethodBuilder.isInherited(psiMethod, srcClass);
         } else {
             overridden = false;
             inherited = false;
         }
-        isInInterface = isInterface(psiMethod);
+        isInInterface = MethodBuilder.isInterface(psiMethod);
         methodId = PsiMethodUtils.formatMethodId(psiMethod);
         accessible = typeDictionary.isAccessible(psiMethod);
-        isSynthetic = isSyntheticMethod(psiMethod);
+        isSynthetic = MethodBuilder.isSyntheticMethod(psiMethod);
         Optional<PsiSubstitutor> methodSubstitutor = findMethodSubstitutor(psiMethod, srcClass, ownerClassPsiType);
-        returnType = resolveReturnType(psiMethod, maxRecursionDepth, typeDictionary, methodSubstitutor);
-        methodParams = extractMethodParams(psiMethod, primaryConstructor, maxRecursionDepth, typeDictionary, methodSubstitutor);
-    }
-
-    static boolean isRelevant(PsiClass psiClass, PsiMethod psiMethod) {
-        boolean isRelevant = true;
-        final PsiClass containingClass = psiMethod.getContainingClass();
-        final PsiClass ownerClass = containingClass == null ? psiClass : containingClass;
-        if (ownerClass != null && isLanguageInherited(ownerClass.getQualifiedName())) {
-            isRelevant = false;
-        } else {
-            final String methodId = PsiMethodUtils.formatMethodId(psiMethod);
-            if (LanguageUtils.isGroovy(psiMethod.getLanguage())
-                    && (psiMethod.getClass().getCanonicalName().contains("GrGdkMethodImpl") || methodId.endsWith(".invokeMethod(java.lang.String,java.lang.Object)") || methodId.endsWith(".getProperty(java.lang.String)") || methodId
-                    .endsWith(".setProperty(java.lang.String,java.lang.Object)"))) {
-                isRelevant = false;
-            }
-            // Need to resolve methods of mocked dependencies from imported libs. consider performance hit..
-            /* else if(ownerClass!=null && ownerClass.getQualifiedName()!=null){
-                JavaPsiFacade facade = JavaPsiFacade.getInstance( ownerClass.getProject());
-                PsiClass[] possibleClasses = facade.findClasses(ownerClass.getQualifiedName(), GlobalSearchScope.projectScope(ownerClass.getProject()));// todo - test with GlobalSearchScope.allScope(ownerClass.getProject()). Alt. skip the check ?
-                if (possibleClasses.length == 0) {
-                    isRelevant = false;
-                }
-            }*/
-        }
-        return isRelevant;
+        returnType = MethodBuilder.resolveReturnType(psiMethod, maxRecursionDepth, typeDictionary, methodSubstitutor);
+        methodParams = MethodBuilder.extractMethodParams(psiMethod, primaryConstructor, maxRecursionDepth, typeDictionary, methodSubstitutor);
     }
 
     /**
@@ -214,62 +178,21 @@ public class Method {
     }
 
     boolean isTestable(){
-        return !isLanguageInherited(ownerClassCanonicalType) && !isSetter() && !isGetter() && !isConstructor() &&((isDefault()|| isProtected() ) && !isInherited() || isPublic()) && !isOverridden() && !isInInterface() && !isAbstract() && !isSynthetic();
+        return !MethodBuilder.isLanguageInherited(ownerClassCanonicalType) && !isSetter() && !isGetter() && !isConstructor() &&((isDefault()|| isProtected() ) && !isInherited() || isPublic()) && !isOverridden() && !isInInterface() && !isAbstract() && !isSynthetic();
     }
 
     void resolveInternalReferences(PsiMethod psiMethod, TypeDictionary typeDictionary) {
-        if (!isLanguageInherited(getOwnerClassCanonicalType())) {
-            resolveCalledMethods(psiMethod, typeDictionary);
-            resolveReferences(psiMethod,typeDictionary);
-            resolveMethodReferences(psiMethod,typeDictionary);
+        if (!MethodBuilder.isLanguageInherited(getOwnerClassCanonicalType())) {
+            this.directMethodCalls.addAll(MethodBuilder.resolveCalledMethods(psiMethod, typeDictionary));
+            this.methodCalls = this.directMethodCalls;
+            this.internalReferences.addAll(MethodBuilder.resolveReferences(psiMethod,typeDictionary));
+            this.methodReferences.addAll(MethodBuilder.resolveMethodReferences(psiMethod,typeDictionary));
         }
     }
 
-    private static boolean isLanguageInherited(String ownerClassCanonicalType) {
-        return "java.lang.Object".equals(ownerClassCanonicalType) || "java.lang.Class".equals(ownerClassCanonicalType) || "groovy.lang.GroovyObjectSupport".equals(ownerClassCanonicalType);
-    }
-
-    @Nullable
-    private Type resolveReturnType(PsiMethod psiMethod, int maxRecursionDepth, TypeDictionary typeDictionary, Optional<PsiSubstitutor> methodSubstitutor) {
-        final PsiType psiType = psiMethod.getReturnType();
-        if (psiType == null) {
-            return null;
-        } else {
-            Optional<PsiType> substitutedType = methodSubstitutor.map(psiSubstitutor -> psiSubstitutor.substitute(psiType));
-            Object typeElement = null;
-            if (LanguageUtils.isScala(psiMethod.getLanguage())) {
-                typeElement = ScalaPsiTreeUtils.resolveReturnType(psiMethod);
-            }
-            return typeDictionary.getType(substitutedType.orElse(psiType), maxRecursionDepth, true,typeElement);
-        }
-    }
-
-    private static PsiField resolveLeftHandExpressionAsField(PsiExpression expr) {
-        PsiElement parent = PsiTreeUtil.skipParentsOfType(expr, PsiParenthesizedExpression.class);
-        if (!(parent instanceof PsiAssignmentExpression)) {
-            return null;
-        }
-        final PsiAssignmentExpression psiAssignmentExpression = (PsiAssignmentExpression) parent;
-        final PsiReference reference = psiAssignmentExpression.getLExpression().getReference();
-        final PsiElement element = reference != null ? reference.resolve() : null;
-        return element == null || !(element instanceof PsiField) ? null : (PsiField)element ;
-    }
-    private boolean isInterface(PsiMethod psiMethod) {
-//            //method inherited from an interface but implemented by this interface should not be considered as interface method
-//            return psiMethod.hasModifierProperty("abstract") || psiMethod.getContainingClass() != null && psiMethod.getContainingClass().isInterface();
-        return psiMethod.hasModifierProperty(PsiModifier.ABSTRACT);
-    }
-
-    private boolean isSyntheticMethod(PsiMethod psiMethod) {
-        if (LanguageUtils.isScala(psiMethod.getLanguage())) {
-            return ScalaPsiTreeUtils.isSyntheticMethod(psiMethod);
-        } else {
-            return false;
-        }
-    }
-    private Optional<PsiSubstitutor> findMethodSubstitutor(PsiMethod psiMethod, PsiClass srcClass, @Nullable  PsiType ownerClassPsiType) {
+    public  Optional<PsiSubstitutor> findMethodSubstitutor(PsiMethod psiMethod, PsiClass srcClass, @Nullable  PsiType ownerClassPsiType) {
 //        if (/*isInherited() && */(isTestable() || isConstructor()) && srcClass != null && hasGenericType(psiMethod)) {
-        if (isInherited() && isTestable() && srcClass != null && hasGenericType(psiMethod)) {
+        if (inherited && isTestable() && srcClass != null && MethodBuilder.hasGenericType(psiMethod)) {
             //todo debug simpler case, check if logic can be simplified without relying to method name first
             List<Pair<PsiMethod, PsiSubstitutor>> methodsSubstitutors = srcClass.findMethodsAndTheirSubstitutorsByName(psiMethod.getName(), true);
             return methodsSubstitutors.stream()
@@ -285,128 +208,6 @@ public class Method {
         }
     }
 
-    private boolean hasGenericType(PsiMethod psiMethod) {
-        return Stream.concat(Stream.of(psiMethod.getParameterList().getParameters()).map(PsiVariable::getType), Stream.of(psiMethod.getReturnType())).anyMatch(this::mayContainTypeParameter);
-    }
-
-    private boolean mayContainTypeParameter(PsiType psiType) {
-        return psiType instanceof PsiClassReferenceType/* && ((PsiClassReferenceType) psiType).resolve() instanceof PsiTypeParameter */;
-    }
-
-    private void resolveReferences(PsiMethod psiMethod, TypeDictionary typeDictionary) {
-        if (LanguageUtils.isGroovy(psiMethod.getLanguage())) {
-            for (ResolvedReference resolvedReference : GroovyPsiTreeUtils.findReferences(psiMethod)) {
-                internalReferences.add(new Reference(resolvedReference.getReferenceName(), resolvedReference.getRefType(), resolvedReference.getPsiOwnerType(), typeDictionary));
-            }
-        }
-        else {
-            for (ResolvedReference resolvedReference : JavaPsiTreeUtils.findReferences(psiMethod)) {
-                internalReferences.add(new Reference(resolvedReference.getReferenceName(), resolvedReference.getRefType(), resolvedReference.getPsiOwnerType(), typeDictionary));
-            }
-        }
-    }
-    private void resolveMethodReferences(PsiMethod psiMethod, TypeDictionary typeDictionary) {
-        for (PsiMethod resolvedMethodReference : JavaPsiTreeUtils.findMethodReferences(psiMethod)) {
-            if (isRelevant(resolvedMethodReference.getContainingClass(), resolvedMethodReference)) {
-                this.methodReferences.add(new Method(resolvedMethodReference, resolvedMethodReference.getContainingClass(), 1, typeDictionary, null));
-            }
-        }
-    }
-    private void resolveCalledMethods(PsiMethod psiMethod, TypeDictionary typeDictionary) {
-        //todo try to pass/support src class in scala/groovy as well. if successful, consider re-implementing with a factory method call
-        if (LanguageUtils.isGroovy(psiMethod.getLanguage())) {
-            for (ResolvedMethodCall resolvedMethodCall : GroovyPsiTreeUtils.findMethodCalls(psiMethod)) {
-                addDirectMethodCallIfRelevant(typeDictionary, resolvedMethodCall, null);
-            }
-        }
-        else if (LanguageUtils.isScala(psiMethod.getLanguage())) {
-            for (ResolvedMethodCall resolvedMethodCall : ScalaPsiTreeUtils.findMethodCalls(psiMethod)) {
-                addDirectMethodCallIfRelevant(typeDictionary, resolvedMethodCall, null);
-            }
-        }
-        else {
-            for (ResolvedMethodCall methodCalled : JavaPsiTreeUtils.findMethodCalls(psiMethod)) {
-                addDirectMethodCallIfRelevant(typeDictionary, methodCalled, methodCalled.getPsiMethod().getContainingClass());
-            }
-        }
-        methodCalls = this.directMethodCalls;
-    }
-
-    private void addDirectMethodCallIfRelevant(TypeDictionary typeDictionary, ResolvedMethodCall methodCalled, PsiClass srcClass) {
-        if (isRelevant(methodCalled.getPsiMethod().getContainingClass(), methodCalled.getPsiMethod())) {
-            this.directMethodCalls.add(new MethodCall(new Method(methodCalled.getPsiMethod(), srcClass, 1, typeDictionary, null),convertArgs(methodCalled.getMethodCallArguments())));
-        }
-    }
-
-    private List<MethodCallArgument> convertArgs(List<MethodCallArg> methodCallArguments) {
-        final ArrayList<MethodCallArgument> methodCallArgs = new ArrayList<>();
-        if (methodCallArguments != null) {
-            for (MethodCallArg methodCallArgument : methodCallArguments) {
-                methodCallArgs.add(new MethodCallArgument(methodCallArgument.getText()));
-            }
-        }
-        return methodCallArgs;
-    }
-
-    private boolean isOverriddenInChild(PsiMethod method, PsiClass srcClass) {
-        String srcQualifiedName = srcClass.getQualifiedName();
-        String methodClsQualifiedName = method.getContainingClass()==null?null:method.getContainingClass().getQualifiedName();
-        if (srcQualifiedName == null || methodClsQualifiedName == null || srcQualifiedName.equals(methodClsQualifiedName)) {
-            return false;
-        }
-        else{
-            final PsiMethod childMethod = MethodSignatureUtil.findMethodBySuperMethod(srcClass, method, false);
-            return childMethod != null;
-        }
-    }
-
-    private boolean isInherited(PsiMethod method, PsiClass srcClass) {
-        String srcQualifiedName = srcClass.getQualifiedName();
-        String methodClsQualifiedName = method.getContainingClass()==null?null:method.getContainingClass().getQualifiedName();
-        return (srcQualifiedName!=null && methodClsQualifiedName!=null &&  !srcQualifiedName.equals(methodClsQualifiedName));
-    }
-
-    private List<Param> extractMethodParams(PsiMethod psiMethod, boolean shouldResolveAllMethods, int maxRecursionDepth, TypeDictionary typeDictionary, Optional<PsiSubstitutor> methodSubstitutor) {
-        ArrayList<Param> params = new ArrayList<>();
-        final PsiParameter[] parameters;
-        if (LanguageUtils.isScala(psiMethod.getLanguage())) {
-            parameters = ScalaPsiTreeUtils.resolveParameters(psiMethod);
-        } else {
-            parameters = psiMethod.getParameterList().getParameters();
-        }
-        for (PsiParameter psiParameter : parameters) {
-            Optional<PsiType> substitutedType = methodSubstitutor.map(psiSubstitutor -> psiSubstitutor.substitute(psiParameter.getType()));
-            final ArrayList<Field> assignedToFields = findMatchingFields(psiParameter, psiMethod);
-            params.add(new Param(psiParameter,substitutedType,typeDictionary,maxRecursionDepth,assignedToFields,shouldResolveAllMethods));
-        }
-        return params;
-    }
-
-    private static ArrayList<Field> findMatchingFields(PsiParameter psiParameter, PsiMethod psiMethod) {
-        final ArrayList<Field> fields = new ArrayList<>();
-        try {
-            if (!psiMethod.hasModifierProperty(PsiModifier.STATIC)) {
-                LocalSearchScope searchScope = LanguageUtils.isScala(psiMethod.getLanguage()) && psiMethod.getContainingClass()!=null ? new LocalSearchScope(psiMethod.getContainingClass()) : new LocalSearchScope(psiMethod);
-                for (PsiReference reference : ReferencesSearch.search(psiParameter, searchScope)) {
-                    final PsiElement element = reference.getElement();
-                    PsiField psiField = null;
-                    if (LanguageUtils.isGroovy(element.getLanguage())) {
-                        psiField = GroovyPsiTreeUtils.resolveGrLeftHandExpressionAsField(element);
-                    }
-                    else if (element instanceof PsiExpression && !PsiUtil.isOnAssignmentLeftHand((PsiExpression) element)) {
-                        psiField = resolveLeftHandExpressionAsField((PsiExpression) element);
-                    }
-                    if (psiField != null && psiField.getContainingClass() != null) {
-                        fields.add(new Field(psiField, psiField.getContainingClass(),null, 0));
-                    }
-                }
-            }
-        } catch (Throwable e) {
-            LOG.warn(String.format("cant search for matching fields for parameter %s in method %s", psiParameter.getName(), psiMethod.getName()),e);
-        }
-        return fields;
-    }
-
     @Override
     public boolean equals(Object o) {
         if (this == o) {
@@ -415,10 +216,7 @@ public class Method {
         if (!(o instanceof Method)) {
             return false;
         }
-
-        Method method = (Method) o;
-
-        return methodId.equals(method.methodId);
+        return methodId.equals(((Method) o).methodId);
     }
 
     @Override
